requirements.txt
fastapi
uvicorn
celery
redis
sqlalchemy
aiomysql
python-multipart
PyMySQL
python-decouple
faker
flower
# eventlet
